do some file refactoring
